requirements.txt
supabase
nltk
gensim
wordcloud
pillow
python-dotenv
